MAINT: refactoring ahead
