+++ enigma-principal/app/src/boot_network/mod.rs
@@ -1,2 +1,3 @@
 pub mod principal_manager;
 pub mod principal_utils;
+pub mod deploy_scripts;
