+++ enigma-principal/app/src/web3_utils/mod.rs
@@ -1,2 +1,1 @@
-pub mod w3utils;
 pub mod deploy_scripts;
